Added features :[authentication,authorization,error handling,validation,file upload and download invoice,pagination logic,stripe API] ;

+++ controllers/admin.js
@@ -2,6 +2,9 @@ const Product = require("../models/product");
 const mongodb = require("mongodb");
 const User = require("../models/user");
 // const ObjectId = new mongodb.ObjectId();
+const { validationResult } = require("express-validator");
+const mongoose = require("mongoose");
+const fileHelper = require("../util/file");
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -11,21 +14,70 @@ exports.getAddProduct = (req, res, next) => {
     productCSS: true,
     activeAddProduct: true,
     editing: false,
+    isAuthenticated: req.session.isLoggedIn,
+    hasError: false,
+    errorMessage: [],
+    validationErrors: [],
   });
 };
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
 
+  if (!image) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "edit Product",
+      path: "admin/edit-product",
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description,
+      },
+      errorMessage: [
+        {
+          msg: "attached file is not an image",
+        },
+      ],
+      isAuthenticated: req.session.isLoggedIn,
+      validationErrors: [],
+    });
+  }
+
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "add Product",
+      path: "admin/add-product",
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        imageUrl: imageUrl,
+        price: price,
+        description: description,
+      },
+      errorMessage: errors.array(),
+      isAuthenticated: req.session.isLoggedIn,
+      validationErrors: errors.array(),
+    });
+  }
+
+  const imageUrl = image.path;
+
   const product = new Product({
+    // _id: new mongoose.Types.ObjectId("6528b172fb6b6333ea1c3e07"),
     title: title,
     price: price,
     description: description,
-    imageUrl: imageUrl,
     userId: req.user,
+    isAuthenticated: req.session.isLoggedIn,
+    imageUrl: imageUrl,
   });
 
   product
@@ -34,33 +86,80 @@ exports.postAddProduct = (req, res, next) => {
       console.log("Created Product");
       res.redirect("/admin/products");
     })
-    .catch((error) => {
-      console.log("error");
+    .catch((err) => {
+      // console.log("errors occured" + error);
+      // return res.status(500).render("admin/edit-product", {
+      //   pageTitle: "edit Product",
+      //   path: "admin/add-product",
+      //   editing: false,
+      //   hasError: true,
+      //   product: {
+      //     title: title,
+      //     imageUrl: imageUrl,
+      //     price: price,
+      //     description: description,
+      //   },
+      //   errorMessage: [{ msg: "Database error" }],
+      //   isAuthenticated: req.session.isLoggedIn,
+      //   validationErrors: [],
+      // });
+      // res.redirect("/500");
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
     });
 };
 
 exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedPrice = req.body.price;
   const updatedDesc = req.body.description;
 
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "edit Product",
+      path: "admin/edit-product",
+      editing: true,
+      hasError: true,
+      product: {
+        title: updatedTitle,
+        imageUrl: image.path,
+        price: updatedPrice,
+        description: updatedDesc,
+      },
+      errorMessage: errors.array(),
+      isAuthenticated: req.session.isLoggedIn,
+      validationErrors: errors.array(),
+    });
+  }
+
   Product.findById(prodId)
     .then((product) => {
+      if (product.userId.toString() !== req.user._id.toString()) {
+        return res.redirect("/");
+      }
       product.title = updatedTitle;
-      product.imageUrl = updatedImageUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       product.description = updatedDesc;
       product.price = updatedPrice;
 
-      product.save();
+      product.save().then((result) => {
+        console.log("product updated");
+        res.redirect("/admin/products");
+      });
     })
-    .then((result) => {
-      console.log("product updated");
-      res.redirect("/admin/products");
-    })
-    .catch((error) => {
-      console.log(error);
+
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
     });
 };
 
@@ -82,21 +181,38 @@ exports.getEditProduct = (req, res, next) => {
         path: "admin/edit-product",
         editing: editMode,
         product: product,
+        isAuthenticated: req.session.isLoggedIn,
+        hasError: false,
+        errorMessage: [],
+        validationErrors: [],
       });
     })
-    .catch((error) => {
-      console.log(error);
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
     });
 };
 
 exports.deleteProduct = (req, res, next) => {
-  const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId })
-    .then()
-    .catch((error) => {
-      console.log(error);
+  const prodId = req.params.productId;
+  Product.findById({ _id: prodId, userId: req.user._id })
+    .then((prod) => {
+      if (!prod || req.user._id.toString() !== prod.userId.toString()) {
+        return next(new Error("Unauthorized"));
+      }
+      fileHelper.deleteFile(prod.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
+    .then(() => {
+      console.log("product deleted");
+      res.status(200).json({
+        message: "Success!",
+      });
+    })
+    .catch((err) => {
+      res.status(500).json();
     });
-  res.redirect("/admin/products");
 };
 
 exports.getProducts = (req, res, next) => {
@@ -107,9 +223,12 @@ exports.getProducts = (req, res, next) => {
         prods: products,
         pageTitle: "Admin Products",
         path: "/admin/products",
+        isAuthenticated: req.session.isLoggedIn,
       });
     })
-    .catch((error) => {
-      console.log(error);
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
     });
 };
